python/requirements.txt
# Python 依赖

# HTTP 请求
requests>=2.31.0

# UI 自动化
playwright>=1.40.0

# 配置文件
pyyaml>=6.0

# 图片处理（视觉测试）
pillow>=10.0.0

# 测试框架（可选）
pytest>=7.4.0
pytest-playwright>=0.4.0
# pytest>=7.4.0
# pytest-playwright>=0.4.0

# 测试报告（待添加）
# allure-pytest>=2.13.0